requirements.txt
numpy>=1.23.4
tqdm
more-itertools
transformers>=4.19.0
ffmpeg-python==0.2.0
click>=8.1.3
PyAudioWPatch
SpeechRecognition
pydub>=0.25.1
git+https://github.com/openai/whisper.git
--extra-index-url https://download.pytorch.org/whl/cu113
torch==1.12.1+cu113
python-osc>=1.8.0
websockets>=10.4
argostranslate>=1.7.0
unidecode>=1.3.6
pykakasi>=2.2.1
ctranslate2
fairseq
sentencepiece
protobuf
progressbar2
